refactor: restructure router and vite module organization
- Update router and vite file structure to improve clarity and maintainability.
- Move router and vite related exports to their respective index files.
- Remove deprecated router and vite files to streamline the codebase.
- Enhance type definitions for plugin integration in vite-related modules.
